Add binary conversion solution
Two passes. Because we iterate from left to right, but the place counter
starts from 0 from right to left, we need to find the length of the
string to know what order the highest order bit is.

First pass counts the string length. We get our place counter here.
Second pass checks if there is a 1 at each iteration, and if so
calculates the decimal value of that place and adds it to the total.
Then we iterate +1 in the byte array and -1 in the place counter.